Move and modularise route state
Route state deserves its own top-level entry, outside of `ui` (which is
mostly just a loose collection of various bits of state, anyway).

This change moves it out of `ui` and modularises it.
