+++ client/landing/login/common.js
@@ -13,7 +13,7 @@ import getSuperProps from 'lib/analytics/super-props';
 import { bindState as bindWpLocaleState } from 'lib/wp/localization';
 import { getUrlParts } from 'lib/url';
 import { setCurrentUser } from 'state/current-user/actions';
-import setRouteAction from 'state/ui/actions/set-route';
+import { setRoute as setRouteAction } from 'state/route/actions';
 
 const debug = debugFactory( 'calypso' );
