+++ client/state/route/init.js
@@ -1,0 +1,7 @@
+/**
+ * Internal dependencies
+ */
+import { registerReducer } from 'state/redux-store';
+import routeReducer from './reducer';
+
+registerReducer( [ 'route' ], routeReducer );
