+++ client/state/route/path/reducer.js
@@ -1,8 +1,4 @@
 /**
- * External dependencies
- */
-
-/**
  * Internal dependencies
  */
 import { ROUTE_SET } from 'state/action-types';
