+++ client/state/selectors/get-current-query-arguments.js
@@ -1,12 +1,12 @@
 /**
  * External dependencies
  */
-
 import { get } from 'lodash';
 
 /**
  * Internal dependencies
  */
+import 'state/route/init';
 
 /**
  * Gets the last query parameters set by a ROUTE_SET action
@@ -14,6 +14,6 @@ import { get } from 'lodash';
  * @param {object} state - global redux state
  * @returns {object} current state value
  */
-export const getCurrentQueryArguments = ( state ) => get( state, 'ui.route.query.current', null );
+export const getCurrentQueryArguments = ( state ) => get( state, 'route.query.current', null );
 
 export default getCurrentQueryArguments;
