+++ client/state/selectors/get-current-route.js
@@ -1,12 +1,12 @@
 /**
  * External dependencies
  */
-
 import { get } from 'lodash';
 
 /**
  * Internal dependencies
  */
+import 'state/route/init';
 
 /**
  * Gets the last route set by a ROUTE_SET action
@@ -14,6 +14,6 @@ import { get } from 'lodash';
  * @param {object} state - global redux state
  * @returns {string} current route value
  */
-export const getCurrentRoute = ( state ) => get( state, 'ui.route.path.current', null );
+export const getCurrentRoute = ( state ) => get( state, 'route.path.current', null );
 
 export default getCurrentRoute;
