+++ client/state/selectors/get-initial-query-arguments.js
@@ -1,12 +1,12 @@
 /**
  * External dependencies
  */
-
 import { get } from 'lodash';
 
 /**
  * Internal dependencies
  */
+import 'state/route/init';
 
 /**
  * Gets the initial query parameters set by a ROUTE_SET action
@@ -14,6 +14,6 @@ import { get } from 'lodash';
  * @param {object} state - global redux state
  * @returns {object} initial query arguments
  */
-export const getInitialQueryArguments = ( state ) => get( state, 'ui.route.query.initial', null );
+export const getInitialQueryArguments = ( state ) => get( state, 'route.query.initial', null );
 
 export default getInitialQueryArguments;
