+++ client/state/selectors/get-previous-path.js
@@ -1,12 +1,12 @@
 /**
  * External dependencies
  */
-
 import { get } from 'lodash';
 
 /**
  * Internal dependencies
  */
+import 'state/route/init';
 
 /**
  * Gets the previous path set by a ROUTE_SET action
@@ -14,6 +14,6 @@ import { get } from 'lodash';
  * @param {object} state - global redux state
  * @returns {string} previous path value
  */
-export const getPreviousPath = ( state ) => get( state, 'ui.route.path.previous', '' );
+export const getPreviousPath = ( state ) => get( state, 'route.path.previous', '' );
 
 export default getPreviousPath;
