+++ client/state/selectors/test/get-partner-slug-from-query.js
@@ -10,12 +10,10 @@ describe( '#getPartnerSlugFromQuery', () => {
 
 	test( 'should return null if no partner_id query present', () => {
 		const state = {
-			ui: {
-				route: {
-					query: {
-						current: {
-							email_address: 'user@wordpress.com',
-						},
+			route: {
+				query: {
+					current: {
+						email_address: 'user@wordpress.com',
 					},
 				},
 			},
@@ -25,13 +23,11 @@ describe( '#getPartnerSlugFromQuery', () => {
 
 	test( 'should return null when partner_id present but not integer', () => {
 		const state = {
-			ui: {
-				route: {
-					query: {
-						current: {
-							email_address: 'user@wordpress.com',
-							partner_id: 'meh',
-						},
+			route: {
+				query: {
+					current: {
+						email_address: 'user@wordpress.com',
+						partner_id: 'meh',
 					},
 				},
 			},
@@ -42,13 +38,11 @@ describe( '#getPartnerSlugFromQuery', () => {
 
 	test( 'should return pressable when partner_id is 49640', () => {
 		const state = {
-			ui: {
-				route: {
-					query: {
-						current: {
-							email_address: 'user@wordpress.com',
-							partner_id: '49640',
-						},
+			route: {
+				query: {
+					current: {
+						email_address: 'user@wordpress.com',
+						partner_id: '49640',
 					},
 				},
 			},
