+++ client/state/ui/reducer.js
@@ -28,7 +28,6 @@ import payment from './payment/reducer';
 import postTypeList from './post-type-list/reducer';
 import preview from './preview/reducer';
 import reader from './reader/reducer';
-import route from './route/reducer';
 import section from './section/reducer';
 import themeSetup from './theme-setup/reducers';
 
@@ -123,7 +122,6 @@ const reducer = combineReducers( {
 	postTypeList,
 	preview,
 	reader,
-	route,
 	section,
 	selectedSiteId,
 	siteSelectionInitialized,
