--- client/state/ui/route/reducer.js
@@ -1,18 +1,0 @@
-/**
- * External dependencies
- *
- */
-
-/**
- * Internal dependencies
- */
-import { combineReducers } from 'state/utils';
-import path from './path/reducer';
-import query from './query/reducer';
-
-const route = combineReducers( {
-	path,
-	query,
-} );
-
-export default route;
